fix(observability): cast health params to int so player counts persist
The CASE expression in MarkHealth had two untyped arms ($3/$4 unbound
params and a bare NULL ELSE), so Postgres resolved the result type to
text and rejected the write into the integer players_online/players_max
columns (SQLSTATE 42804). Cast the params to ::int so the CASE result
matches the columns.

+++ internal/repository/game_server.go
@@ -257,8 +257,8 @@ func (r *GameServerRepository) UnassignHost(ctx context.Context, id string) erro
 func (r *GameServerRepository) MarkHealth(ctx context.Context, id string, reachable bool, online, max int) error {
 	_, err := r.pool.Exec(ctx, `
 		UPDATE game_servers
-		SET players_online = CASE WHEN $2 THEN $3 ELSE NULL END,
-		    players_max    = CASE WHEN $2 THEN $4 ELSE NULL END,
+		SET players_online = CASE WHEN $2 THEN $3::int ELSE NULL END,
+		    players_max    = CASE WHEN $2 THEN $4::int ELSE NULL END,
 		    last_seen      = CASE WHEN $2 THEN now() ELSE last_seen END
 		WHERE id = $1 AND deleted_at IS NULL AND status = 'running'`,
 		id, reachable, online, max)
